lib: sbi: Support sbi_hart_switch_mode() for Zcheripurecap ABI
- Enable CRE in menvfg CSR to enable CHERI accessing in supervisor
  mode when it has Zcherihybrid extension.
- Delegate cheri exception to supervisor mode.
- Pass capability arguments for sbi_hart_switch_mode().
- Capability register cleanup to prevent capability leakage when mode
  switching.

+++ include/sbi/sbi_hart.h
@@ -148,8 +148,8 @@ bool sbi_hart_has_csr(struct sbi_scratch *scratch, enum sbi_hart_csrs csr);
 void __attribute__((noreturn)) sbi_hart_hang(void);
 
 void __attribute__((noreturn))
-sbi_hart_switch_mode(unsigned long arg0, unsigned long arg1,
-		     unsigned long next_addr, unsigned long next_mode,
+sbi_hart_switch_mode(uintptr_t arg0, uintptr_t arg1,
+		     uintptr_t next_addr, unsigned long next_mode,
 		     bool next_virt);
 
 #endif
